Implement like dislike trackactivities

+++ internal/service/memberships/login.go
@@ -26,7 +26,7 @@ func (s *service) Login(request memberships.LoginRequest) (string, error) {
 		fmt.Println("test case password not match")
 		return "", errors.New("email and password not mathced")
 	}
-	accessToken, err := jwt.CreateToken(int64(userDetail.ID), userDetail.Username, s.cfg.Service.SecretJWT)
+	accessToken, err := jwt.CreateToken(userDetail.ID, userDetail.Username, s.cfg.Service.SecretJWT)
 
 	if err != nil {
 		log.Error().Err(err).Msg("failed to create jwt token")
