+++ pkg/jwt/jwt.go
@@ -7,7 +7,7 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 )
 
-func CreateToken(id int64, username, secretKey string) (string, error) {
+func CreateToken(id uint, username, secretKey string) (string, error) {
 
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
 		"id":       id,
@@ -23,7 +23,7 @@ func CreateToken(id int64, username, secretKey string) (string, error) {
 	return tokenStr, nil
 }
 
-func ValidateToken(tokenStr, secretkey string) (int64, string, error) {
+func ValidateToken(tokenStr, secretkey string) (uint, string, error) {
 	key := []byte(secretkey)
 	claims := jwt.MapClaims{}
 
@@ -36,7 +36,7 @@ func ValidateToken(tokenStr, secretkey string) (int64, string, error) {
 	if !token.Valid {
 		return 0, "", errors.New("invalid token")
 	}
-	return int64(claims["id"].(float64)), claims["username"].(string), nil
+	return uint(claims["id"].(float64)), claims["username"].(string), nil
 }
 
 func ValidateTokenWithoutExpiry(tokenStr, secretkey string) (int64, string, error) {
